server/requirements.txt
torch
torchvision
flask
flask-cors
numpy
pillow
Werkzeug==2.2.2
fastapi